pyproject.toml
[tool.poetry]
name = "crm-bank-system"
version = "0.1.0"
description = ""
authors = ["fed0r <wisp47344@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.10"
celery = "^5.4.0"
requests = "^2.32.3"
django-celery-beat = "^2.7.0"
python-docx = "^1.1.2"
openpyxl = "^3.1.5"
mongoengine = "^0.29.1"
pymongo = "^4.10.1"
plotly = "^5.24.1"
quickchart-io = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
